Added the Getting Started section to the documentation
JAVA-1701
