+++ docs/reference/content/getting-started/index.md
@@ -1,0 +1,15 @@
++++
+date = "2015-03-17T15:36:56Z"
+title = "Getting Started"
+[menu.main]
+  weight = 10
+  pre = "<i class='fa fa-road'></i>"
++++
+
+## Getting Started
+
+To help you get started quickly on the new driver, follow:
+
+  * [Installation]({{< ref "getting-started/installation-guide.md" >}})
+  * [Quick Tour]({{< ref "getting-started/quick-tour.md" >}})
+  * [Admin Quick Tour]({{< ref "getting-started/quick-tour-admin.md" >}})
