--- docs/reference/content/tutorial/admin-operations.md
@@ -1,11 +1,0 @@
-+++
-date = "2015-03-17T15:36:56Z"
-draft = true
-title = "Admin Operations"
-[menu.main]
-  parent = "Getting Started"
-  weight = 20
-  pre = "<i class='fa'></i>"
-+++
-
-# Admin Introduction
