--- docs/reference/content/tutorial/index.md
@@ -1,15 +1,0 @@
-+++
-date = "2015-03-17T15:36:56Z"
-title = "Getting Started"
-[menu.main]
-  weight = 10
-  pre = "<i class='fa fa-road'></i>"
-+++
-
-## Getting Started
-
-To help you get started quickly on the new driver follow:
-
-  * [Installation]({{< ref "tutorial/installation-guide.md" >}})
-  * [CRUD Operations]({{< ref "tutorial/crud-operations.md" >}})
-  * [Admin Operations]({{< ref "tutorial/admin-operations.md" >}})
