+++ docs/reference/layouts/shortcodes/apiref.html
@@ -1,1 +1,1 @@
-{{ .Page.Site.Data.mongodb.apiUrl }}?{{ .Get 0 }}.html
+{{ .Page.Site.Data.mongodb.apiUrl }}?{{ with (.Get 0) }}{{ . }}{{ if ne (in . ".html") true}}.html{{ end }}{{ end }}
