+++ docs/reference/themes/mongodb/layouts/shortcodes/note.html
@@ -1,4 +1,4 @@
 <div class="admonition {{if .Get "class"}}{{.Get "class"}}{{else}}note{{end}}">
-<h5 class="admonition-title">Note</h5>
+<h5 class="admonition-title">{{if .Get "class"}}{{.Get "class"}}{{else}}Note{{end}}</h5>
 {{ .Inner }}
 </div>
